Checkin before I do drastic things

+++ Parallel Computing/Lab6/mandelbrot_omp.h
@@ -3,10 +3,10 @@
 
 #include "int_array.h"
 
-extern double mandelbrot_width;
-extern double mandelbrot_height;
+extern float  mandelbrot_width;
+extern float  mandelbrot_height;
 extern int    mandelbrot_iters;
 
-void mandelbrot_omp(int x_min, int x_max, int y_min, int y_max, double mag, IntArray arr);
+void mandelbrot_omp(int x_min, int x_max, int y_min, int y_max, float mag, IntArray arr);
 
 #endif
